Fixed plugin importer
