+++ Dockerfile
@@ -10,7 +10,7 @@ WORKDIR /app
 
 COPY . /app
 
-RUN pip3 --no-cache-dir install -r requirements.txt
+RUN pip3 install -r requirements.txt
 
 RUN yarn
