+++ docker-compose.yml
@@ -3,5 +3,7 @@ services:
   bot:
     image: peachbot/peach:latest
     build: .
+    ports: 
+     - "5000:5000"
     volumes:
       - ./:/app
