+++ peach_bot/source/eventhandler.py
@@ -1,0 +1,21 @@
+class EventHandler:
+    
+    async def on_message(self, message):
+        self.log.info("Received message: {0}#{1}@{2} --> {3}".format(message.author.name, message.author.discriminator, message.guild.name, message.content))
+
+        #ignore messages sent by the bot
+        if message.author == self.user:
+            return
+        #filter for manual page invokes
+        if message.content.startswith('!man'):
+            await self.pluginhandler.man(message)
+
+        #try to run a command in message starts with prefix
+        elif message.content.startswith('!'):
+            await self.pluginhandler.runcommand(message)
+
+
+    async def on_member_join(self, member):
+        # Welcome message
+        await member.guild.system_channel.send('{0.mention} felt cute.'.format(member))
+        self.log.info('{0.mention} joined {0.guild.name}'.format(member))
